Modules Github, chat_info and user_info migrated to KanTeX and dropped utils.tgdoc
